Switch from BedrockSkinUtility to ViaBedrockUtility for skins.
